requirements.txt
Flask==2.3.2
biopython==1.81
numpy==1.24.3
matplotlib==3.7.1